requirements.txt
requests==2.25.0
python-dotenv==0.15.0
instabot==0.117.0
